Change activity to take a double instead of a BigDecimal.
It makes more sense to use a double on the outside when creating a debit
or credit and leave the mechanics of using monetary amounts under the
hood away from the main parts of the application.

+++ src/main/java/org/vbc4me/awanna/account/Session.java
@@ -86,7 +86,7 @@ public class Session {
 	 * @param date  of this activity
 	 * @param activity for members to participate in
 	 */
-	public void addActivity(LocalDate date, LocalTime time, String activity, BigDecimal cost) {
+	public void addActivity(LocalDate date, LocalTime time, String activity, double cost) {
 		checkIfSessionIsOpen();
 		Activity act = Activity.build().date(date).time(time).activity(activity).cost(cost).create();
 		activities.put(act.date(), act);
